stuff/pyproject.toml
[tool.poetry]
name = "casestudy3"
version = "0.1.0"
description = "Crap for campus numerique"
authors = ["Your Name <you@example.com>"]
license = "for money"

[tool.poetry.dependencies]
python = "3.8.12"
Flask = "^2.0.3"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
